Changed the Petri net datastructure to work with a map for the flows.
Instead of storing the flows as triples (from, to, weight), we now
store the flows as a map: (P x T) U (T x P) -> N. The Petri net model
has been edited to take these changes into account, but only does so
when retrieving the Petri net. For storing the Petri net this is still
a todo-item.

The new datastructure does not allow an immediate translation to JSON,
as the map that is used cannot be converted to a JSON object that mimics
its behaviour. Instead, the map has to be converted back to an array of
triples. A converter has been written - PetrinetToJson - that facilitates
this functionality.