Changed the Lola to Petri net converter and translator for the new Petri net ds
The LolaToPetrinet converter and the PetrinetTranslator converter have been
changed to account for the changes of the Petri net data structure. A bug fix
regarding the setting of the flows for a particular Petri net was needed and is
also part of this commit. Furthermore, the Petri net model is now able to succesfully
store Petri nets in the database.