Added new exception type for the server.
This new exception type includes an http status code that can then be used
by the error handler to return a response with a correct status code.
As a consequence, the showError method of the Controller class was deemed
no longer to be needed, and therefore removed. All appropriate instances
of exceptions have been converted to the new exception type.

This commit also includes two bug fixes: when a user would upload a file
to the server that would not be accepted, the user folder would retain
some files about the user. This should not happen, and this is now fixed.
The other fix checks whether the set of places and the set of transitions
of a Petri net are disjoint. If they are not disjoint an exception is thrown.