Started implementation of the new Logger.
This new logger splits the management of session from the user meta data.
For each user a log file is created containing meta data regarding the sessions
of the user. This file includes a counter for the amount of sessions this user
has undertaken. This counter can then be used as a pointer for the the current
session.

Every session is a seperate file. Each session describes all the graphs the
user went through in their quest towards a solution. When a user wants to model
a graph for a different Petri net, a new session is started, meaning that a new
file is created belonging to this session and the logs are then filed in this file.

The old logger stored all this information in a single file. This is simple, but also
very taxing on the file system when many sessions are started by the user. Because JSON
encoding is used for each log, the entire log needs to be read and parsed whenever a
change is made. When a lot of data is stored in a log this can be very taxing. The new
logger solves this by using a seperate file for each session. While we end up with more
files, the size of the files will be shorter, and therefore reading and writing will be
done more quickly.

The new logger only supports the creation of the meta log file for a specific user.
Functions for creation and management of session log file are being worked on.