Refactoring of the Logger.
The logger has been refactored to follow MVC architecture. The logger
is now a model, with the SessionController acting as the Controller. To fit
this role, the logger has been renamed to SessionModel.

Because not every model needs a connection to a database, like the SessionModel,
the base class for models has been modified, or rather renamed. Every model
that now interacts with a database inherits from DatabaseModel.

Furthermore, the namespacing for the QueryBuilder has been changed.

Note that the appendGraph method of the SessionModel has not yet been tested,
and is pretty much guaranteed to go wrong at this point.