Refactored Petri net and graph datastructures for the refactor
Both data structures have been simplified. Methods that gave
the same result but with different wrappers have been merged
to one function. For example, the preset function could return
the places in the preset of a transition of a Petri net, or it
could also include the flows. This method now always returns flows
as you can easily retrieve the place from the flow.

Models and converters have been adapted to take these changes into
account.