Coverability graph checker seems to work correctly.
All markings now have a property $reachable, detailing whether
the marking is reachable from the initial marking. The BreadthFirstSearch
class had to be adapted for the changes in the graph data structure.

The coverability graph checker requires set operations that are not provided
by the standard set class, therefore a SetUtils class has been created.