Added new Feedback Code for omitting Omega while it occurs in the preset
When the graph contains a marking for which at least one of the places
does not omega tokens assigned to it, while at least one of the markings
in its preset does assign to those places omega, then the markin is incorrect:
A place cannot suddenly become bounded again.

The client has been adapted to include feedback message translations and style
changes when the student/user makes this kind of error.