Removed addedOmega requirement
This requirement is enforced implicitly as it is impossible to still
arrive at a valid marking without adding an omega when the second
if-level is entered. The requirement that the unbounded set is a
subset of the coverable set is enough to make sure that no incorrect
markings get marked as correct.