Created a more OOP Petrinet architecture
Since PHP does not support generics, the only way (as far as I'm
aware) to enforce type safety in regards as to what elements are
stored in a set or map, is to create a wrapper around these
datastructures which contain functions that modify the underlying
datastructure, but can only receive arguments of a certain type. These
classes make this possible. Furthermore, to separate construction
logic from semantics for both markings and Petri nets, there are now
builder classes for both. Note that in order to create a new marking
you still require access to the Petri net for which it is being
defined. Otherwise we could end up with markings that assign tokens to
places that do not exist in the Petri net.