Attempted error handling in the handlers, with some success
Though I am not happy with how for each exception type, you need to
redo the same few steps. It would be much nicer if a handler could
specify for each exception type which http-status code it should give
to the response. I will try to find a method for that later. For now,
I think it would be a better idea to move all the files to their
domain categories