When requesting or checking a net a marking can be included
When requesting a Petri net you can add a marking id in the query
URL. CoRA will then try to include that marking for you in the
response. For this some extra views had to be created. When it comes
to providing feedback, an initial marking is _required_. Currently the
initial marking id is provided via a query string in the URL, but this
will most likely be changed to be a part of the JSON request format.