The user id for registering a petrinet should be part of the body
When registering a Petri net the user id should be supplied in the
body of the request. No longer should you need to specify the user id
in the URL. Additionally, fixed a bug in the RegisterPetrinet service
where it would throw a non-existing exception, which would then also
not be caught. Furthermore a BadRequest exception has been added. This
exception can be thrown whenever a request is malformed. The invoke
method of the abstract request handler has been adapted to catch these
exceptions and then return appropriate responses.