Better routes for sessions
Session details should be encoded in the body of the request. One
caveat: when getting the current session, since we're using a GET
request, and PHP does not parse form-data in GET requests, only in
POST requests, and Slim does not provide any sort of parsing, using
application/form-data is not valid when getting the current session of
a user.