A Different way of getting feedback via sessions
The session log has been extended so that a session also includes the
initial marking of the Petri net. This makes it possible that a user
only needs to specify its user id and session id when requesting
feedback. All other information can be retrieved from the log. This
has been implemented.